+++ src/main/java/bl4ckscor3/mod/theplopper/Configuration.java
@@ -4,14 +4,12 @@ import org.apache.commons.lang3.tuple.Pair;
 
 import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.common.ForgeConfigSpec.BooleanValue;
-import net.minecraftforge.common.ForgeConfigSpec.IntValue;
 
 public class Configuration
 {
 	public static final ForgeConfigSpec CONFIG_SPEC;
 	public static final Configuration CONFIG;
 
-	public final IntValue range;
 	public final BooleanValue displayParticles;
 	public final BooleanValue playSound;
 	public final BooleanValue bypassOutputSide;
@@ -26,9 +24,6 @@ public class Configuration
 
 	Configuration(ForgeConfigSpec.Builder builder)
 	{
-		range = builder
-				.comment("The range in blocks at which the plopper will still pick up items")
-				.defineInRange("range", 3, 1, 20);
 		displayParticles = builder
 				.comment("Whether particles should show up every time the plopper picks up an item")
 				.define("displayParticles", true);