+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-# v1.1
+# v1.2
+- Added range upgrade
+- Removed range config option
+- Fix tooltips not showing up in The Plopper's GUI
 
+# v1.1
 - Fix items stacking in the plopper's inventory when they had a different metadata but the same item ID
 
-
 # v1.0
-
 - Initial release